celery-requirements.txt
flask==1.1.4
flask-cors==3.0.10 
psycopg2==2.9.3 
SQLAlchemy==1.4.39 
Flask-SQLAlchemy==2.5.1 
bcrypt==3.2.2 
flask-jwt-extended==3.25.1
zenpy==2.0.24 
datadog==0.44.0 
Werkzeug==1.0.1
Flask-Caching==2.0.0 
sentry-sdk==1.7.1 
Flask-Restx==0.5.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
numpy==1.23.0 
pandas==1.4.3
boto3==1.28.55
pyjwt==1.7.1 
nose2==0.11.0 
aws_xray_sdk==2.10.0 
coverage==6.4.1 
sendgrid==6.9.7 
itsdangerous==1.1.0
pyOpenSSL==22.0.0 
Flask-HTTPAuth==4.7.0 
cryptography==37.0.2 
croniter==1.3.5 
pyotp==2.6.0 
pyqrcode==1.2.1 
pypng==0.0.21 
blinker==1.4 
MarkupSafe==2.0.1
Pillow==9.2.0
requests==2.28.2
oauthlib==3.2.1
celery==5.1.2
click==7.1.1
firebase-admin==6.0.0
Flask-Pydantic==0.11.0
Flask-Marshmallow==0.15.0
eventlet==0.33.3
Redis==5.0.1
python-dotenv
ffmpeg-python
openai-whisper
ffprobe-python
openai==1.6.0
transformers==4.36.2
python-docx
uwsgi
